[refactor] 변경사항 저장 모달의 범용성을 생각해서 common으로 분리
